requirements.txt
torch
torchvision
numpy
Pillow
openai
requests
